running test with nose
