+++ publish/tests/example_app/views.py
@@ -1,0 +1,1 @@
+# Create your views here.
